Extract `TypeCastingTest` into `test/cases/quoting_test.rb`
