+++ activerecord/test/cases/adapters/postgresql/quoting_test.rb
@@ -1,5 +1,4 @@
 require "cases/helper"
-require "ipaddr"
 
 module ActiveRecord
   module ConnectionAdapters
