+++ activerecord/test/cases/quoting_test.rb
@@ -150,6 +150,62 @@ module ActiveRecord
       end
     end
 
+    class TypeCastingTest < ActiveRecord::TestCase
+      def setup
+        @conn = ActiveRecord::Base.connection
+      end
+
+      def test_type_cast_symbol
+        assert_equal "foo", @conn.type_cast(:foo)
+      end
+
+      def test_type_cast_date
+        date = Date.today
+        expected = @conn.quoted_date(date)
+        assert_equal expected, @conn.type_cast(date)
+      end
+
+      def test_type_cast_time
+        time = Time.now
+        expected = @conn.quoted_date(time)
+        assert_equal expected, @conn.type_cast(time)
+      end
+
+      def test_type_cast_numeric
+        assert_equal 10, @conn.type_cast(10)
+        assert_equal 2.2, @conn.type_cast(2.2)
+      end
+
+      def test_type_cast_nil
+        assert_nil @conn.type_cast(nil)
+      end
+
+      def test_type_cast_unknown_should_raise_error
+        obj = Class.new.new
+        assert_raise(TypeError) { @conn.type_cast(obj) }
+      end
+
+      def test_type_cast_object_which_responds_to_quoted_id
+        quoted_id_obj = Class.new {
+          def quoted_id
+            "'zomg'"
+          end
+
+          def id
+            10
+          end
+        }.new
+        assert_equal 10, @conn.type_cast(quoted_id_obj)
+
+        quoted_id_obj = Class.new {
+          def quoted_id
+            "'zomg'"
+          end
+        }.new
+        assert_raise(TypeError) { @conn.type_cast(quoted_id_obj) }
+      end
+    end
+
     class QuoteBooleanTest < ActiveRecord::TestCase
       def setup
         @connection = ActiveRecord::Base.connection
